requirements.txt
streamlit>=1.20.0
requests>=2.28.0
python-dotenv>=1.0.0
SQLAlchemy>=1.4
email-validator>=1.3.1
